+++ train_gpt2.py
@@ -26,17 +26,9 @@ from dataclasses import dataclass
 from pathlib import Path
 
 from typing import (
-    AbstractSet,
-    cast,
-    Collection,
-    Dict,
-    Iterator,
     List,
-    Literal,
     Optional,
-    Sequence,
     Tuple,
-    Union,
 )
 
 import numpy as np
@@ -48,216 +40,13 @@ from torch.nn.parallel import DistributedDataParallel as DDP
 from torch.distributed import init_process_group, destroy_process_group
 from torch.distributed.optim import ZeroRedundancyOptimizer
 import torch.distributed as dist
-from tiktoken.load import load_tiktoken_bpe
+
+from llmc_py.tokenizer import Tokenizer
+from llmc_py.rope import precompute_freqs_cis, apply_rotary_emb
+from llmc_py.utils import repeat_kv, sample_top_p, RMSNorm
 
 # -----------------------------------------------------------------------------
 # PyTorch nn.Module definitions for the GPT-2 model
-
-# The tiktoken tokenizer can handle <=400k chars without
-# pyo3_runtime.PanicException.
-TIKTOKEN_MAX_ENCODE_CHARS = 400_000
-
-# https://github.com/openai/tiktoken/issues/195
-# Here we iterate over subsequences and split if we exceed the limit
-# of max consecutive non-whitespace or whitespace characters.
-MAX_NO_WHITESPACES_CHARS = 25_000
-
-
-class Tokenizer:
-    """
-    Tokenizing and encoding/decoding text using the Tiktoken tokenizer.
-    """
-
-    special_tokens: Dict[str, int]
-
-    num_reserved_special_tokens = 256
-
-    pat_str = r"(?i:'s|'t|'re|'ve|'m|'ll|'d)|[^\r\n\p{L}\p{N}]?\p{L}+|\p{N}{1,3}| ?[^\s\p{L}\p{N}]+[\r\n]*|\s*[\r\n]+|\s+(?!\S)|\s+"  # noqa: E501
-
-    def __init__(self, model_path: str):
-        """
-        Initializes the Tokenizer with a Tiktoken model.
-
-        Args:
-            model_path (str): The path to the Tiktoken model file.
-        """
-        assert os.path.isfile(model_path), model_path
-
-        mergeable_ranks = load_tiktoken_bpe(model_path)
-        num_base_tokens = len(mergeable_ranks)
-        special_tokens = [
-            "<|begin_of_text|>",
-            "<|end_of_text|>",
-            "<|reserved_special_token_0|>",
-            "<|reserved_special_token_1|>",
-            "<|finetune_right_pad_id|>",
-            "<|step_id|>",
-            "<|start_header_id|>",
-            "<|end_header_id|>",
-            "<|eom_id|>",  # end of message
-            "<|eot_id|>",  # end of turn
-            "<|python_tag|>",
-        ]
-        reserved_tokens = [
-            f"<|reserved_special_token_{2 + i}|>"
-            for i in range(self.num_reserved_special_tokens - len(special_tokens))
-        ]
-        special_tokens = special_tokens + reserved_tokens
-
-        self.special_tokens = {
-            token: num_base_tokens + i for i, token in enumerate(special_tokens)
-        }
-        self.model = tiktoken.Encoding(
-            name=Path(model_path).name,
-            pat_str=self.pat_str,
-            mergeable_ranks=mergeable_ranks,
-            special_tokens=self.special_tokens,
-        )
-
-        self.n_words: int = num_base_tokens + len(special_tokens)
-        # BOS / EOS token IDs
-        self.bos_id: int = self.special_tokens["<|begin_of_text|>"]
-        self.eos_id: int = self.special_tokens["<|end_of_text|>"]
-        self.eot_id: int = self.special_tokens["<|eot_id|>"]
-        self.eom_id: int = self.special_tokens["<|eom_id|>"]
-        self.python_tag_id = self.special_tokens["<|python_tag|>"]
-        self.pad_id: int = self.special_tokens["<|finetune_right_pad_id|>"]
-        self.stop_tokens = [
-            self.special_tokens["<|eom_id|>"],
-            self.special_tokens["<|eot_id|>"],
-        ]
-
-    def encode(
-        self,
-        s: str,
-        *,
-        bos: bool,
-        eos: bool,
-        allowed_special: Optional[Union[Literal["all"], AbstractSet[str]]] = None,
-        disallowed_special: Union[Literal["all"], Collection[str]] = (),
-    ) -> List[int]:
-        """
-        Encodes a string into a list of token IDs.
-
-        Args:
-            s (str): The input string to be encoded.
-            bos (bool): Whether to prepend the beginning-of-sequence token.
-            eos (bool): Whether to append the end-of-sequence token.
-            allowed_tokens ("all"|set[str]): allowed special tokens in string
-            disallowed_tokens ("all"|set[str]): special tokens that raise an error when in string
-
-        Returns:
-            list[int]: A list of token IDs.
-
-        By default, setting disallowed_special=() encodes a string by ignoring
-        special tokens. Specifically:
-        - Setting `disallowed_special` to () will cause all text corresponding
-          to special tokens to be encoded as natural text (insteading of raising
-          an error).
-        - Setting `allowed_special` to "all" will treat all text corresponding
-          to special tokens to be encoded as special tokens.
-        """
-        if allowed_special is None:
-            allowed_special = set()
-        assert type(s) is str
-
-        substrs = (
-            substr
-            for i in range(0, len(s), TIKTOKEN_MAX_ENCODE_CHARS)
-            for substr in self._split_whitespaces_or_nonwhitespaces(
-                s[i : i + TIKTOKEN_MAX_ENCODE_CHARS], MAX_NO_WHITESPACES_CHARS
-            )
-        )
-        t: List[int] = []
-        for substr in substrs:
-            t.extend(
-                self.model.encode(
-                    substr,
-                    allowed_special=allowed_special,
-                    disallowed_special=disallowed_special,
-                )
-            )
-        if bos:
-            t.insert(0, self.bos_id)
-        if eos:
-            t.append(self.eos_id)
-        return t
-
-    def decode(self, t: Sequence[int]) -> str:
-        """
-        Decodes a list of token IDs into a string.
-
-        Args:
-            t (List[int]): The list of token IDs to be decoded.
-
-        Returns:
-            str: The decoded string.
-        """
-        # Typecast is safe here. Tiktoken doesn't do anything list-related with the sequence.
-        return self.model.decode(cast(List[int], t))
-
-    @staticmethod
-    def _split_whitespaces_or_nonwhitespaces(
-        s: str, max_consecutive_slice_len: int
-    ) -> Iterator[str]:
-        """
-        Splits the string `s` so that each substring contains no more than `max_consecutive_slice_len`
-        consecutive whitespaces or consecutive non-whitespaces.
-        """
-        current_slice_len = 0
-        current_slice_is_space = s[0].isspace() if len(s) > 0 else False
-        slice_start = 0
-
-        for i in range(len(s)):
-            is_now_space = s[i].isspace()
-
-            if current_slice_is_space ^ is_now_space:
-                current_slice_len = 1
-                current_slice_is_space = is_now_space
-            else:
-                current_slice_len += 1
-                if current_slice_len > max_consecutive_slice_len:
-                    yield s[slice_start:i]
-                    slice_start = i
-                    current_slice_len = 1
-        yield s[slice_start:]
-
-class RMSNorm(torch.nn.Module):
-    def __init__(self, dim: int, eps: float = 1e-6):
-        super().__init__()
-        self.eps = eps
-        self.weight = nn.Parameter(torch.ones(dim))
-
-    def _norm(self, x):
-        return x * torch.rsqrt(x.pow(2).mean(-1, keepdim=True) + self.eps)
-
-    def forward(self, x):
-        output = self._norm(x.float()).type_as(x)
-        return output * self.weight
-
-def sample_top_p(probs, p):
-    """
-    Perform top-p (nucleus) sampling on a probability distribution.
-
-    Args:
-        probs (torch.Tensor): Probability distribution tensor.
-        p (float): Probability threshold for top-p sampling.
-
-    Returns:
-        torch.Tensor: Sampled token indices.
-
-    Note:
-        Top-p sampling selects the smallest set of tokens whose cumulative probability mass
-        exceeds the threshold p. The distribution is renormalized based on the selected tokens.
-    """
-    probs_sort, probs_idx = torch.sort(probs, dim=-1, descending=True)
-    probs_sum = torch.cumsum(probs_sort, dim=-1)
-    mask = probs_sum - probs_sort > p
-    probs_sort[mask] = 0.0
-    probs_sort.div_(probs_sort.sum(dim=-1, keepdim=True))
-    next_token = torch.multinomial(probs_sort, num_samples=1)
-    next_token = torch.gather(probs_idx, -1, next_token)
-    return next_token
 
 class NewGELU(nn.Module):
     """Careful there are a few versions of GeLU, this one is the exact one used by OpenAI"""
@@ -267,79 +56,13 @@ class NewGELU(nn.Module):
 # using a global to toggle flash-attention
 FLASH = 0
 
-def reshape_for_broadcast(freqs_cis: torch.Tensor, x: torch.Tensor):
-    ndim = x.ndim
-    assert 0 <= 1 < ndim
-    assert freqs_cis.shape == (x.shape[1], x.shape[-1])
-    shape = [d if i == 1 or i == ndim - 1 else 1 for i, d in enumerate(x.shape)]
-    return freqs_cis.view(*shape)
-
-def apply_rotary_emb(
-    xq: torch.Tensor,
-    xk: torch.Tensor,
-    freqs_cis: torch.Tensor,
-) -> Tuple[torch.Tensor, torch.Tensor]:
-    xq_ = torch.view_as_complex(xq.float().reshape(*xq.shape[:-1], -1, 2))
-    xk_ = torch.view_as_complex(xk.float().reshape(*xk.shape[:-1], -1, 2))
-    freqs_cis = reshape_for_broadcast(freqs_cis, xq_)
-    xq_out = torch.view_as_real(xq_ * freqs_cis).flatten(3)
-    xk_out = torch.view_as_real(xk_ * freqs_cis).flatten(3)
-    return xq_out.type_as(xq), xk_out.type_as(xk)
-
-def repeat_kv(x: torch.Tensor, n_rep: int) -> torch.Tensor:
-    """torch.repeat_interleave(x, dim=2, repeats=n_rep)"""
-    bs, slen, n_kv_heads, head_dim = x.shape
-    if n_rep == 1:
-        return x
-    return (
-        x[:, :, :, None, :]
-        .expand(bs, slen, n_kv_heads, n_rep, head_dim)
-        .reshape(bs, slen, n_kv_heads * n_rep, head_dim)
-    )
-
-def apply_scaling(freqs: torch.Tensor):
-    # Values obtained from grid search
-    scale_factor = 8
-    low_freq_factor = 1
-    high_freq_factor = 4
-    old_context_len = 8192  # original llama3 length
-
-    low_freq_wavelen = old_context_len / low_freq_factor
-    high_freq_wavelen = old_context_len / high_freq_factor
-    new_freqs = []
-    for freq in freqs:
-        wavelen = 2 * math.pi / freq
-        if wavelen < high_freq_wavelen:
-            new_freqs.append(freq)
-        elif wavelen > low_freq_wavelen:
-            new_freqs.append(freq / scale_factor)
-        else:
-            assert low_freq_wavelen != high_freq_wavelen
-            smooth = (old_context_len / wavelen - low_freq_factor) / (
-                high_freq_factor - low_freq_factor
-            )
-            new_freqs.append((1 - smooth) * freq / scale_factor + smooth * freq)
-    return torch.tensor(new_freqs, dtype=freqs.dtype, device=freqs.device)
-
-def precompute_freqs_cis(
-    dim: int, end: int, theta: float = 10000.0, use_scaled: bool = False
-):
-    freqs = 1.0 / (theta ** (torch.arange(0, dim, 2)[: (dim // 2)].float() / dim))
-    t = torch.arange(end, device=freqs.device, dtype=torch.float32)
-    if use_scaled:
-        freqs = apply_scaling(freqs)
-    freqs = torch.outer(t, freqs)
-    freqs_cis = torch.polar(torch.ones_like(freqs), freqs)  # complex64
-    return freqs_cis
-
 class CausalSelfAttention(nn.Module):
 
     def __init__(self, config):
         super().__init__()
         assert config.n_embd % config.n_head == 0
-        is_llama = config.is_llama
-        self.is_llama = is_llama
-        if not is_llama:
+        self.is_llama = config.is_llama
+        if not self.is_llama:
             # key, query, value projections for all heads, but in a batch
             self.c_attn = nn.Linear(config.n_embd, 3 * config.n_embd)
             # output projection
@@ -349,20 +72,20 @@ class CausalSelfAttention(nn.Module):
             self.n_head = config.n_head
             self.n_embd = config.n_embd
         else:
-            self.n_kv_heads = config.n_kv_head
-            self.n_local_heads = config.n_head
-            self.n_local_kv_heads = self.n_kv_heads
-            self.n_rep = self.n_local_heads // self.n_local_kv_heads
+            self.n_head = config.n_head
+            self.n_kv_head = config.n_kv_head
+            self.n_rep = self.n_head // self.n_kv_head
             self.head_dim = config.n_embd // config.n_head
 
+            # TODO(gordicaleksa): this can be easily made the same as the above (c_attn, c_proj)
             self.wq = nn.Linear(config.n_embd, config.n_head * self.head_dim, bias=False)
-            self.wk = nn.Linear(config.n_embd, self.n_kv_heads * self.head_dim, bias=False)
-            self.wv = nn.Linear(config.n_embd, self.n_kv_heads * self.head_dim, bias=False)
+            self.wk = nn.Linear(config.n_embd, config.n_kv_head * self.head_dim, bias=False)
+            self.wv = nn.Linear(config.n_embd, config.n_kv_head * self.head_dim, bias=False)
             self.wo = nn.Linear(config.n_head * self.head_dim, config.n_embd, bias=False)
 
         # not really a 'bias', more of a mask, but following the OpenAI/HF naming though
         self.register_buffer("bias", torch.tril(torch.ones(config.block_size, config.block_size))
-                                    .view(1, 1, config.block_size, config.block_size))
+                                     .view(1, 1, config.block_size, config.block_size))
 
     def forward(self, x, freqs_cis=None):
         if not self.is_llama:
@@ -388,13 +111,14 @@ class CausalSelfAttention(nn.Module):
             y = self.c_proj(y)
             return y
         else:
+            # TODO(gordicaleksa): this can be easily merged with the if branch above
             bsz, seqlen, _ = x.shape
 
             # QKV
             xq, xk, xv = self.wq(x), self.wk(x), self.wv(x)
-            xq = xq.view(bsz, seqlen, self.n_local_heads, self.head_dim)
-            xk = xk.view(bsz, seqlen, self.n_local_kv_heads, self.head_dim)
-            xv = xv.view(bsz, seqlen, self.n_local_kv_heads, self.head_dim)
+            xq = xq.view(bsz, seqlen, self.n_head, self.head_dim)
+            xk = xk.view(bsz, seqlen, self.n_kv_head, self.head_dim)
+            xv = xv.view(bsz, seqlen, self.n_kv_head, self.head_dim)
             # rotate QK (rope)
             xq, xk = apply_rotary_emb(xq, xk, freqs_cis=freqs_cis)
 
@@ -419,9 +143,8 @@ class MLP(nn.Module):
 
     def __init__(self, config):
         super().__init__()
-        is_llama = config.is_llama
-        self.is_llama = is_llama
-        if not is_llama:
+        self.is_llama = config.is_llama
+        if not self.is_llama:
             self.c_fc    = nn.Linear(config.n_embd, 4 * config.n_embd)
             self.gelu    = NewGELU()
             self.c_proj  = nn.Linear(4 * config.n_embd, config.n_embd)
@@ -450,10 +173,9 @@ class Block(nn.Module):
 
     def __init__(self, config):
         super().__init__()
-        is_llama = config.is_llama
-        self.ln_1 = RMSNorm(config.n_embd, config.norm_eps) if is_llama else nn.LayerNorm(config.n_embd)
+        self.ln_1 = RMSNorm(config.n_embd, config.norm_eps) if config.is_llama else nn.LayerNorm(config.n_embd)
         self.attn = CausalSelfAttention(config)
-        self.ln_2 = RMSNorm(config.n_embd, config.norm_eps) if is_llama else nn.LayerNorm(config.n_embd)
+        self.ln_2 = RMSNorm(config.n_embd, config.norm_eps) if config.is_llama else nn.LayerNorm(config.n_embd)
         self.mlp = MLP(config)
 
     def forward(self, x, freqs_cis=None):
@@ -474,8 +196,9 @@ class GPTConfig:
     n_embd: int = 768
 
 @dataclass
-class Llama31Config:
+class LlamaConfig:
     is_llama = True
+    version: str = "3.1"
     block_size: int = 1024
     vocab_size: int = 128256
     n_layer: int = 32
@@ -493,27 +216,26 @@ class GPT(nn.Module):
     def __init__(self, config):
         super().__init__()
         self.config = config
-        is_llama = config.is_llama
-        self.is_llama = is_llama
+        self.is_llama = config.is_llama
 
         self.transformer = nn.ModuleDict(dict(
             wte = nn.Embedding(config.vocab_size, config.n_embd),
-            **({'wpe': nn.Embedding(config.block_size, config.n_embd)} if not is_llama else {}),
+            **({} if self.is_llama else {'wpe': nn.Embedding(config.block_size, config.n_embd)}),
             h = nn.ModuleList([Block(config) for _ in range(config.n_layer)]),
-            ln_f = RMSNorm(config.n_embd, config.norm_eps) if is_llama else nn.LayerNorm(config.n_embd),
+            ln_f = RMSNorm(config.n_embd, config.norm_eps) if self.is_llama else nn.LayerNorm(config.n_embd),
         ))
         self.lm_head = nn.Linear(config.n_embd, config.vocab_size, bias=False)
         self.lm_head.LLMC_SKIP_INIT = 1 # don't init this one, we will tie weights
-        if not is_llama:
+        if not self.is_llama:
             self.transformer.wte.weight = self.lm_head.weight # https://paperswithcode.com/method/weight-tying
 
         # init all weights, use a torch rng object to be very careful
         self.init_rng = torch.Generator()
         self.init_rng.manual_seed(42)
-        if not is_llama:
+        if not self.is_llama:
             self.apply(self._init_weights)
 
-        if is_llama:
+        if self.is_llama:
             self.freqs_cis = precompute_freqs_cis(
                 config.n_embd // config.n_head,
                 config.block_size * 2,
@@ -556,11 +278,15 @@ class GPT(nn.Module):
 
         if targets is not None:
             # if we are given some desired targets also calculate the loss
-            logits = self.lm_head(x).float()
+            logits = self.lm_head(x)
+            if self.is_llama:
+                logits = logits.float()
             loss = F.cross_entropy(logits.view(-1, logits.size(-1)), targets.view(-1), ignore_index=-1)
         else:
             # inference-time mini-optimization: only forward the lm_head on the very last position
-            logits = self.lm_head(x[:, [-1], :]).float() # note: using list [-1] to preserve the time dim
+            logits = self.lm_head(x[:, [-1], :]) # note: using list [-1] to preserve the time dim
+            if self.is_llama:
+                logits = logits.float()
             loss = None
 
         # there are performance reasons why not returning logits is prudent, if not needed
@@ -570,13 +296,12 @@ class GPT(nn.Module):
         return logits, loss
 
     @staticmethod
-    def modify_llama_keys(checkpoint, config: Llama31Config):
-        # 1) rename key tok_embeddings.weight to transformer.wte.weight
+    def adapt_llama_state_dict_keys(checkpoint, config: LlamaConfig):
+        # rename key tok_embeddings.weight to transformer.wte.weight
         checkpoint['transformer.wte.weight'] = checkpoint.pop('tok_embeddings.weight')
 
-        # layers.0.attention_norm.weight -> transformer.h.0.ln_1.weight
-        # layers.0.ffn_norm.weight -> transformer.h.0.ln_2.weight
-        # loop over all layers
+        # layers.x.attention_norm.weight -> transformer.h.x.ln_1.weight
+        # layers.x.ffn_norm.weight -> transformer.h.x.ln_2.weight
         for i in range(config.n_layer):
             for name in ['attention_norm', 'ffn_norm']:
                 for suffix in ['weight']:
@@ -584,11 +309,10 @@ class GPT(nn.Module):
                     new_key = f'transformer.h.{i}.ln_{1 if name == "attention_norm" else 2}.{suffix}'
                     checkpoint[new_key] = checkpoint.pop(old_key)
 
-        # layers.0.attention.wq.weight -> transformer.h.0.attn.wq.weight
-        # layers.0.attention.wk.weight -> transformer.h.0.attn.wk.weight
-        # layers.0.attention.wv.weight -> transformer.h.0.attn.wv.weight
-        # layers.0.attention.wo.weight -> transformer.h.0.attn.wo.weight
-        # loop over all layers
+        # layers.x.attention.wq.weight -> transformer.h.x.attn.wq.weight
+        # layers.x.attention.wk.weight -> transformer.h.x.attn.wk.weight
+        # layers.x.attention.wv.weight -> transformer.h.x.attn.wv.weight
+        # layers.x.attention.wo.weight -> transformer.h.x.attn.wo.weight
         for i in range(config.n_layer):
             for name in ['attention.wq', 'attention.wk', 'attention.wv', 'attention.wo']:
                 for suffix in ['weight']:
@@ -596,10 +320,9 @@ class GPT(nn.Module):
                     new_key = f'transformer.h.{i}.attn.{name.split(".")[-1]}.{suffix}'
                     checkpoint[new_key] = checkpoint.pop(old_key)
 
-        # layers.0.feed_forward.w1.weight -> transformer.h.0.mlp.w1.weight
-        # layers.0.feed_forward.w2.weight -> transformer.h.0.mlp.w2.weight
-        # layers.0.feed_forward.w3.weight -> transformer.h.0.mlp.w3.weight
-        # loop over all layers
+        # layers.x.feed_forward.w1.weight -> transformer.h.x.mlp.w1.weight
+        # layers.x.feed_forward.w2.weight -> transformer.h.x.mlp.w2.weight
+        # layers.x.feed_forward.w3.weight -> transformer.h.x.mlp.w3.weight
         for i in range(config.n_layer):
             for name in ['feed_forward.w1', 'feed_forward.w2', 'feed_forward.w3']:
                 for suffix in ['weight']:
@@ -615,16 +338,19 @@ class GPT(nn.Module):
         return checkpoint
 
     @classmethod
-    def from_pretrained_llama3_1(cls):
-        ckpt_dir = "/home/aleksa/Documents/eureka/nano-llama31/llama-models/models/llama3_1/Meta-Llama-3.1-8B"
+    def from_pretrained_llama3(cls, ckpt_dir, tokenizer_path):
+        model_args = LlamaConfig()
+
         ckpt_path = sorted(Path(ckpt_dir).glob("*.pth"))[0]
         checkpoint = torch.load(ckpt_path, map_location="cpu", weights_only=True)
-        checkpoint = GPT.modify_llama_keys(checkpoint, Llama31Config())
-        model_args = Llama31Config()
+        checkpoint = GPT.adapt_llama_state_dict_keys(checkpoint, model_args)
 
         torch.set_default_tensor_type(torch.cuda.BFloat16Tensor)
         model = GPT(model_args)
         model.load_state_dict(checkpoint, strict=False)
+
+        tokenizer = Tokenizer(model_path=tokenizer_path)
+        model.tokenizer = tokenizer
         return model
 
     @classmethod
@@ -735,7 +461,7 @@ class GPT(nn.Module):
         return idx
 
     @torch.inference_mode()
-    def generate2(
+    def generate_llama(
         self,
         prompt_tokens: List[List[int]],
         max_gen_len: int,
@@ -1089,7 +815,9 @@ if __name__ == "__main__":
     # default settings will overfit a tiny batch of data
     # and save model weights and debug state to disk on the first iteration
     parser = argparse.ArgumentParser()
-    parser.add_argument("--use_llama3", type=int, default=1, help="use llama3 model")
+    parser.add_argument("--llama3", type=int, default=1, help="use llama3 model")
+    parser.add_argument("--llama3_ckpt_dir", type=str, default=None, help="path to llama3 model checkpoint")
+    parser.add_argument("--llama3_tokenizer_path", type=str, default=None, help="path to llama3 tokenizer")
     # file system input / output
     parser.add_argument("--input_bin", type=str, default="dev/data/tinyshakespeare/tiny_shakespeare_val.bin", help="input .bin to train on")
     parser.add_argument("--input_val_bin", type=str, default="", help="input .bin to eval validation loss on")
@@ -1176,7 +904,7 @@ if __name__ == "__main__":
 
     # set up a context manager following the desired dtype and device
     ptdtype = {'float32': torch.float32, 'bfloat16': torch.bfloat16, 'float16': torch.float16}[args.dtype]
-    ctx = torch.amp.autocast(device_type=device_type, dtype=ptdtype) if (device_type == "cuda" and not args.use_llama3) else nullcontext()
+    ctx = torch.amp.autocast(device_type=device_type, dtype=ptdtype) if (device_type == "cuda" and not args.llama3) else nullcontext()
 
     # rng / reproducibility
     torch.manual_seed(42)
@@ -1208,8 +936,10 @@ if __name__ == "__main__":
         }[args.model]
         model = GPT(model_config)
     else:
-        if args.use_llama3:
-            model = GPT.from_pretrained_llama3_1()
+        if args.llama3:
+            assert args.llama3_ckpt_dir is not None and os.path.exists(args.llama3_ckpt_dir), f"llama3 ckpt dir {args.llama3_ckpt_dir} does not exist"
+            assert args.llama3_tokenizer_path is not None and os.path.exists(args.llama3_tokenizer_path), f"llama3 tokenizer path {args.llama3_tokenizer_path} does not exist"
+            model = GPT.from_pretrained_llama3(args.llama3_ckpt_dir, args.llama3_tokenizer_path)
         else:
             # load the GPT-2 model weights
             model = GPT.from_pretrained(args.model)
@@ -1232,10 +962,37 @@ if __name__ == "__main__":
         val_loader = DistributedDataLoader(args.input_val_bin, B, T, ddp_rank, ddp_world_size)
 
     # -------------------------------------------------------------------------
+    # LLaMA 3 inference
+    if args.llama3:
+        model.eval()
+        prompts: List[str] = [
+        # For these prompts, the expected answer is the natural continuation of the prompt
+        "Clearly, the meaning of life is",
+        "Simply put, the theory of relativity states that",
+        """The repo llm.c on GitHub is""",
+        # Few shot prompt (providing a few examples before asking model to complete more);
+        """Translate English to French:
+
+        sea otter => loutre de mer
+        peppermint => menthe poivrée
+        plush girafe => girafe peluche
+        cheese =>""",
+            ]
+
+        prompt_tokens = [model.tokenizer.encode(x, bos=True, eos=False) for x in prompts]
+
+        generation_tokens, _ = model.generate_llama(prompt_tokens, max_gen_len=64, temperature=0.6, top_p=0.9, logprobs=False, echo=False)
+        results = [{"generation": model.tokenizer.decode(t)} for t in generation_tokens]
+        for prompt, result in zip(prompts, results):
+            print(prompt, end="")
+            print(f"{result['generation']}")
+            print("\n==================================\n")
+
+    # -------------------------------------------------------------------------
     # PyTorch -> C bridge: save some weights and state for C to load later as reference
 
     # do one forward pass to generate ground truth for our C tests
-    if not args.use_llama3 and master_process and args.write_tensors and (not args.inference_only):
+    if master_process and args.write_tensors and (not args.inference_only):
         x, y = train_loader.next_batch()
         x, y = x.to(device), y.to(device)
         logits, loss = model(x, y)
@@ -1298,63 +1055,41 @@ if __name__ == "__main__":
         last_step = (step == args.num_iterations)
 
         # once in a while evaluate the validation dataset
-        # if (args.val_loss_every > 0 \
-        #     and (step % args.val_loss_every == 0 or last_step)) \
-        #     and (val_loader is not None):
-        #     model.eval()
-        #     val_loader.reset()
-        #     with torch.no_grad():
-        #         val_loss = 0.0
-        #         for _ in range(args.val_max_steps):
-        #             x, y = val_loader.next_batch()
-        #             x, y = x.to(device), y.to(device)
-        #             _, loss = model(x, y, return_logits=False)
-        #             val_loss += loss.item()
-        #         val_loss /= args.val_max_steps
-        #     # log to console and to file
-        #     print0(f"val loss {val_loss}")
-        #     if master_process and logfile is not None:
-        #         with open(logfile, "a") as f:
-        #             f.write("s:%d tel:%f\n" % (step, val_loss))
+        if (args.val_loss_every > 0 \
+            and (step % args.val_loss_every == 0 or last_step)) \
+            and (val_loader is not None):
+            model.eval()
+            val_loader.reset()
+            with torch.no_grad():
+                val_loss = 0.0
+                for _ in range(args.val_max_steps):
+                    x, y = val_loader.next_batch()
+                    x, y = x.to(device), y.to(device)
+                    _, loss = model(x, y, return_logits=False)
+                    val_loss += loss.item()
+                val_loss /= args.val_max_steps
+            # log to console and to file
+            print0(f"val loss {val_loss}")
+            if master_process and logfile is not None:
+                with open(logfile, "a") as f:
+                    f.write("s:%d tel:%f\n" % (step, val_loss))
 
         # once in a while perform model inference on the master process
-        if True:
-            # (args.sample_every > 0 \
-            # and (step % args.sample_every == 0 or last_step)) \
-            # and master_process:
+        if (args.sample_every > 0 \
+            and (step % args.sample_every == 0 or last_step)) \
+            and master_process:
             model.eval()
             # before we end, let's also do one round of inference
             # we'll kick off the generation with "<|endoftext|>", which designates the start of a new sequence
-            # start_ids = [enc.eot_token]
-            # xg = (torch.tensor(start_ids, dtype=torch.long, device=device)[None, ...])
-            # max_new_tokens = 32
-            # temperature = 1.0
-            # top_k = 40
-            tokenizer_path = "/home/aleksa/Documents/eureka/nano-llama31/llama-models/models/llama3_1/Meta-Llama-3.1-8B/tokenizer.model"
-            tokenizer = Tokenizer(model_path=tokenizer_path)
-            raw_model.tokenizer = tokenizer
-            prompts: List[str] = [
-        # For these prompts, the expected answer is the natural continuation of the prompt
-        "Clearly, the meaning of life is",
-        "Simply put, the theory of relativity states that",
-        """The repo llm.c on GitHub is""",
-        # Few shot prompt (providing a few examples before asking model to complete more);
-        """Translate English to French:
-
-        sea otter => loutre de mer
-        peppermint => menthe poivrée
-        plush girafe => girafe peluche
-        cheese =>""",
-            ]
-
-            prompt_tokens = [tokenizer.encode(x, bos=True, eos=False) for x in prompts]
-
-            generation_tokens, _ = raw_model.generate2(prompt_tokens, max_gen_len=64, temperature=0.6, top_p=0.9, logprobs=False, echo=False)
-            results = [{"generation": tokenizer.decode(t)} for t in generation_tokens]
-            for prompt, result in zip(prompts, results):
-                print(prompt, end="") # AK: change end="\n" to end=""
-                print(f"{result['generation']}")
-                print("\n==================================\n")
+            start_ids = [enc.eot_token]
+            xg = (torch.tensor(start_ids, dtype=torch.long, device=device)[None, ...])
+            max_new_tokens = 32
+            temperature = 1.0
+            top_k = 40
+            yg = raw_model.generate(xg, max_new_tokens, temperature=temperature, top_k=top_k)
+            print0('---------------')
+            print0(enc.decode(yg[0].tolist()))
+            print0('---------------')
 
         # bit confusing: we want to make sure to eval and sample on 0th iteration
         # but also after the very last iteration. so we loop for step <= num_iterations
